refactor(api): give POST /runner/tasks/claim a domain module
The handler held one Serializable transaction of ~390 lines wrapped in a
hand-rolled six-attempt retry loop. Both were closure-independent; the
four-line prologue that is not — the runner note, the ownership assertion,
the reconcile and the archived-run notice — stays at the route.

`claimRun(db, { body, claimantClass, now })` owns the transaction, its
isolation level, and the retry. The isolation level and the retry policy are
invariants of claiming, not of HTTP: a caller that restated them would be
restating the reason this transaction can lose a race it did not cause. The
retry loop now has a test, which was impossible while it was welded to a
route.

`PinnedRunTargetError`, `isCandidateActivationFailure` and
`namedFailureReason` had no other caller and move with it.
`openMergeTailStopNotice` is shared with `handleRegressionCompletion`, which
stays in `app.ts`, so it moves to `merge-tail-actions.ts` rather than into
either caller — importing it back out of `app.ts` would make a cycle.
